# Backend/public/api-docs/models/openapi.yaml
openapi: 3.1.1
info:
  title: '🏛️ TDW - ACiencia REST api'
  description: '[UPM] TDW REST api ACiencia'
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  version: 1.8.0
  contact:
    name: Máster Universitario en Ingeniería Web
    url: 'https://miw.etsisi.upm.es/'
    email: miw.etsisi@upm.es
servers:
  - url: '/api/v1'
    description: TDW Production server (uses live data)
...
paths:
...
  '/entities/{entityId}/associations':
    parameters:
      - $ref: '#/components/parameters/entityId'
    get:
      tags: [ '🏢 Entities' ]
      summary: List of associations related to the entity
      description: Returns the associations linked to the entity `entityId`.
      operationId: readEntityAssociations
      responses:
        '200':
          description: Array of Associations
          content:
            application/json:
              schema:
                type: object
                properties:
                  associations:
                    type: array
                    items: { $ref: '#/components/schemas/Association' }
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
        '404': { $ref: '#/components/responses/Response_404' }
...
  '/associations':
    get:
      tags: [ '🤝 Associations' ]
      summary: Retrieves the collection of Association resources.
      operationId: readAssociations
      parameters:
        - $ref: '#/components/parameters/elementTextParam'
        - $ref: '#/components/parameters/orderParam'
        - $ref: '#/components/parameters/orderingParam'
      responses:
        '200':
          description: Object containing an array of Associations
          content:
            application/json:
              schema:
                type: object
                properties:
                  associations:
                    type: array
                    items: { $ref: '#/components/schemas/Association' }
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
        '404': { $ref: '#/components/responses/Response_404' }

    post:
      tags: [ '🤝 Associations' ]
      summary: Creates an Association resource.
      operationId: createAssociation
      security: [ { MiWApiSecurity: [ writer ] } ]
      requestBody:
        $ref: '#/components/requestBodies/AssociationRequestBody'
      responses:
        '201':
          description: '`Created`: association created'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Association' }
          headers:
            Location: { $ref: '#/components/headers/Location' }
        '400': { $ref: '#/components/responses/Response_400' }
        '401': { $ref: '#/components/responses/Response_401' }
        '403': { $ref: '#/components/responses/Response_403' }
        '422': { $ref: '#/components/responses/Response_422' }

    options:
      tags: [ '🤝 Associations' ]
      summary: Provides the list of HTTP supported methods.
      operationId: optionsAssociations
      responses:
        '204':
          description: '`Allow` header <Response body is empty>'
          headers:
            Allow: { $ref: '#/components/headers/Allow' }
  '/associations/{associationId}':
    parameters:
      - $ref: '#/components/parameters/associationId'
    get:
      tags: [ '🤝 Associations' ]
      summary: Retrieves an Association resource based on a single ID.
      operationId: readAssociation
      responses:
        '200':
          description: Association
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Association' }
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
        '404': { $ref: '#/components/responses/Response_404' }

    put:
      tags: [ '🤝 Associations' ]
      summary: Updates the Association resource.
      operationId: updateAssociation
      security: [ { MiWApiSecurity: [ writer ] } ]
      requestBody:
        $ref: '#/components/requestBodies/AssociationRequestBody'
      parameters:
        - $ref: '#/components/parameters/ifMatchParam'
      responses:
        '209':
          description: '`Content Returned`: association updated'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Association' }
        '400': { $ref: '#/components/responses/Response_400' }
        '401': { $ref: '#/components/responses/Response_401' }
        '404': { $ref: '#/components/responses/Response_404' }
        '428': { $ref: '#/components/responses/Response_428' }

    delete:
      tags: [ '🤝 Associations' ]
      summary: Removes the Association resource.
      operationId: deleteAssociation
      security: [ { MiWApiSecurity: [ writer ] } ]
      responses:
        '204':
          description: Association deleted <Response body is empty>
        '401': { $ref: '#/components/responses/Response_401' }
        '404': { $ref: '#/components/responses/Response_404' }
...
components:
  schemas:
...
    Element:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Element Id
          readOnly: true
        name:
          type: string
          maxLength: 80
        birthDate:
          description: An RFC3339 formatted date string ("2025-01-01")
          type: string
          format: date
        deathDate:
          description: An RFC3339 formatted date string ("2025-01-01")
          type: string
          format: date
        imageUrl:
          type: string
          maxLength: 2047
          format: url
        wikiUrl:
          type: string
          maxLength: 2047
          format: url
      required:
        - name
...
    Association:
      allOf:
        - $ref: '#/components/schemas/Element'
        - type: object
          required: [ website ]
          properties:
            website:
              description: Main website URL of the association
              type: string
              format: url
              maxLength: 2047
            entities:
              description: Array of **Entity** IDs related to this association
              type: array
              items:
                type: integer
      examples:
        - association:
            id: 300
            name: Alliance 300
            website: 'https://alliance300.example.com'
            birthDate: '1999-05-07'
            imageUrl: 'https://example.com/alliance.png'
            wikiUrl: 'https://es.wikipedia.org/wiki/Alliance_300'
            entities: [ 2, 42 ]
    Message:
      type: object
      description: HTTP response message
      properties:
        code:
          description: Response code
          type: integer
          format: int64
        message:
          description: Response message
          type: string
  responses:
...
    Response_400:
      description: '`BAD REQUEST`: User name or e-mail already exists, or role does not exist'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-400:
              $ref: '#/components/examples/response-400'
    Response_401:
      description: '`UNAUTHORIZED`: invalid Authorization header'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-401:
              $ref: '#/components/examples/response-401'
    Response_403:
      description: '`FORBIDDEN`: You don''t have permission to access'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-403:
              $ref: '#/components/examples/response-403'
    Response_404:
      description: '`NOT FOUND`: resource not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-404:
              $ref: '#/components/examples/response-404'
...
    Response_428:
      description: '`PRECONDITION REQUIRED`: This request is required to be conditional; try using "If-Match".'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-428:
              $ref: '#/components/examples/response-428'
    Response_422:
      description: '`UNPROCESSABLE ENTITY`: Username, e-mail or password is left out'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-422:
              $ref: '#/components/examples/response-422'
  requestBodies:
    AssociationRequestBody:
      description: '`Association` data'
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              name:
                type: string
                maxLength: 80
              website:
                type: string
                format: url
                maxLength: 2047
              birthDate:
                type: string
                format: date
              deathDate:
                type: string
                format: date
              imageUrl:
                type: string
                format: url
                maxLength: 2047
              wikiUrl:
                type: string
                format: url
                maxLength: 2047
            required: [ 'name', 'website' ]
...
  parameters:
...
    entityId:
      name: entityId
      in: path
      description: ID of entity
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
        pattern: ^\d+$
...
    elementTextParam:
      name: name
      in: query
      description: Elements containing the text
      required: false
      schema:
        type: string
        pattern: '^[a-zA-Z0-9()áéíóúÁÉÍÓÚñÑ %$\.+-<>]+$'
...
    associationId:
      name: associationId
      in: path
      description: ID of association
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
        pattern: ^\d+$
...
    orderParam:
      name: order
      in: query
      description: '`id` | `name`'
      required: false
      allowEmptyValue: false
      schema:
        type: string
        enum:
          - id
          - name
    orderingParam:
      name: ordering
      in: query
      description: '`ASC` | `DESC`'
      required: false
      allowEmptyValue: false
      schema:
        type: string
        enum:
          - ASC
          - DESC
    ifMatchParam:
      name: If-Match
      in: header
      description: "ETag value of the resource you are trying to update"
      required: true
      schema:
        type: string
        pattern: "[0-9a-f]*"
...
  headers:
    ETag:
      description: Entity Tag (unique identifier provided for cache validation)
      schema:
        type: string
        pattern: "[0-9a-f]*"
    Location:
      description: '`URI` of the new resource'
      schema:
        type: string
        format: url
    Allow:
      description: List of HTTP supported methods
      schema:
        type: string
  examples:
    response-400:
      value:
        code: 400
        message: 'BAD REQUEST: name or e-mail already exists'
    response-401:
      value:
        code: 401
        message: 'UNAUTHORIZED: invalid Authorization header'
    response-403:
      value:
        code: 403
        message: "FORBIDDEN: You don't have permission to access"
    response-404:
      value:
        code: 404
        message: "NOT FOUND: Resource not found (or you don't have permission to access)"
...
    response-428:
      value:
        code: 428
        message: 'PRECONDITION REQUIRED: This request is required to be conditional; try using "If-Match".'
    response-422:
      value:
        code: 422
        message: 'UNPROCESSABLE ENTITY: name, e-mail or password is left out'
  securitySchemes:
    MiWApiSecurity:
      type: oauth2
      description: 'This API uses OAuth 2 with the implicit grant flow. [More info](https://api.example.com/docs/auth)'
      bearerFormat: JWT
      flows:
        password:
          tokenUrl: /access_token
          scopes:
            reader: 'Permissions: (read, update) [ User, Person, Entity, Product ]'
            writer: 'can also: (create, update, delete) [ User, Person, Entity, Product ]'